clean up - remove test code not longer to be used
